# swagger.yaml
openapi: 3.0.3
info:
  title: Music Playlist Management System API
  description: |
    A comprehensive API for managing music playlists with Spotify integration.
    
    This API allows users to:
    - Register and authenticate securely
    - Create and manage personal playlists
    - Search for songs using Spotify's extensive catalog
    - Add songs from Spotify to custom playlists
    - Organize and manage their music collection
    
    ## Authentication
    This API uses JWT (JSON Web Tokens) for authentication. After registering or logging in, 
    include the JWT token in the Authorization header as: `Bearer <token>`
    
    ## Getting Started
    1. Register a new account using `/api/auth/register`
    2. Login to get your JWT token using `/api/auth/login`
    3. Start creating playlists and adding songs!
    
    ## Spotify Integration
    The API integrates with Spotify Web API to provide access to millions of songs.
    Make sure your Spotify API credentials are properly configured on the server.
  version: 1.0.0
  contact:
    name: API Support
    email: support@playlisthub.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://your-production-domain.com/api
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /playlists:
    get:
      tags:
        - Playlists
      summary: Get user playlists
      description: Retrieve all playlists for the authenticated user with pagination
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Items per page
        - in: query
          name: search
          schema:
            type: string
          description: Search query for playlist names or descriptions
      responses:
        '200':
          description: Playlists retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Playlist'
                      pagination:
                        $ref: '#/components/schemas/PaginationInfo'

    post:
      tags:
        - Playlists
      summary: Create playlist
      description: Create a new playlist for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaylistRequest'
            example:
              name: My Favorite Songs
              description: A collection of my all-time favorite tracks
      responses:
        '201':
          description: Playlist created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Playlist'
        '400':
          $ref: '#/components/responses/BadRequestError'
...
  /songs:
    get:
      tags:
        - Songs
      summary: Get all songs
      description: Retrieve all songs with pagination and search
      security: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: search
          schema:
            type: string
          description: Search in song name, artist, or album
      responses:
        '200':
          description: Songs retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Song'
                      pagination:
                        $ref: '#/components/schemas/PaginationInfo'

    post:
      tags:
        - Songs
      summary: Create song
      description: Add a new song to the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSongRequest'
      responses:
        '201':
          description: Song created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Song'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
          description: User ID
        username:
          type: string
          description: Unique username
        email:
          type: string
          format: email
          description: User email address
        name:
          type: string
          description: User full name
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        _id: "507f1f77bcf86cd799439011"
        username: "johndoe"
        email: "john@example.com"
        name: "John Doe"
        createdAt: "2023-01-01T00:00:00.000Z"
        updatedAt: "2023-01-01T00:00:00.000Z"
...
    Playlist:
      type: object
      properties:
        _id:
          type: string
          description: Playlist ID
        name:
          type: string
          description: Playlist name
        description:
          type: string
          description: Playlist description
        owner:
          $ref: '#/components/schemas/User'
        songs:
          type: array
          items:
            $ref: '#/components/schemas/Song'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        _id: "507f1f77bcf86cd799439012"
        name: "My Favorite Songs"
        description: "A collection of my all-time favorites"
        owner:
          _id: "507f1f77bcf86cd799439011"
          name: "John Doe"
          username: "johndoe"
        songs: []
        createdAt: "2023-01-01T00:00:00.000Z"
        updatedAt: "2023-01-01T00:00:00.000Z"
    CreatePlaylistRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
...
    Song:
      type: object
      properties:
        _id:
          type: string
          description: Song ID
        spotifyId:
          type: string
          description: Spotify track ID
        name:
          type: string
          description: Song title
        artist:
          type: string
          description: Artist name(s)
        album:
          type: string
          description: Album name
        duration:
          type: integer
          description: Duration in milliseconds
        imageUrl:
          type: string
          format: uri
          description: Album cover image URL
        previewUrl:
          type: string
          format: uri
          description: 30-second preview URL
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        _id: "507f1f77bcf86cd799439013"
        spotifyId: "4uLU6hMCjMI75M1A2tKUQC"
        name: "Blinding Lights"
        artist: "The Weeknd"
        album: "After Hours"
        duration: 200040
        imageUrl: "https://i.scdn.co/image/ab67616d0000b2738863bc11d2aa12b54f5aeb36"
        previewUrl: "https://p.scdn.co/mp3-preview/..."
        createdAt: "2023-01-01T00:00:00.000Z"
        updatedAt: "2023-01-01T00:00:00.000Z"
    CreateSongRequest:
      type: object
      required:
        - spotifyId
        - name
        - artist
        - album
        - duration
      properties:
        spotifyId:
          type: string
          minLength: 22
          maxLength: 22
        name:
          type: string
          minLength: 1
          maxLength: 200
        artist:
          type: string
          minLength: 1
          maxLength: 200
        album:
          type: string
          minLength: 1
          maxLength: 200
        duration:
          type: integer
          minimum: 1000
          maximum: 3600000
        imageUrl:
          type: string
          format: uri
        previewUrl:
          type: string
          format: uri

    # Spotify Schemas
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          description: Success message
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
          description: Detailed validation errors (if applicable)
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        total:
          type: integer
          description: Total number of items
        totalPages:
          type: integer
          description: Total number of pages
      example:
        page: 1
        limit: 20
        total: 100
        totalPages: 5
  responses:
...
    BadRequestError:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Invalid input data"
            details:
              - field: "email"
                message: "Invalid email format"
...